+++ src/main/kotlin/com/keyskey/ktknowledge/infrastructure/database/Users.kt
@@ -1,4 +1,4 @@
-package com.keyskey.ktknowledge.database
+package com.keyskey.ktknowledge.infrastructure.database
 
 import org.ktorm.schema.Table
 import org.ktorm.schema.datetime
